Add `mergeDeepRight.js` and `mergeDeepLeft.js`
