+++ lib/mean/index.js
@@ -1,4 +1,4 @@
-var sum = require('../sum')
+import sum from '../sum'
 
 export default function mean(nums) {
   return sum(nums) / nums.length
