+++ lib/mergeDeepRight/index.js
@@ -1,6 +1,4 @@
 import _curry2 from '../_internal/_curry2'
-import mergeDeepLeft from '../mergeDeepLeft'
+import mergeDeepRight from './mergeDeepRight'
 
-export default _curry2(function mergeDeepRight(first, second) {
-  return mergeDeepLeft(second, first)
-})
+export default _curry2(mergeDeepRight)
