+++ lib/mergeDeepRight/mergeDeepRight.js
@@ -1,0 +1,5 @@
+import mergeDeepLeft from '../mergeDeepLeft/mergeDeepLeft'
+
+export default function mergeDeepRight(first, second) {
+  return mergeDeepLeft(second, first)
+}
